fix reconnect and enable test
