+++ chain/network/src/peer_manager/tests/tier2.rs
@@ -1,18 +1,16 @@
 use crate::network_protocol::testonly as data;
 use crate::peer_manager::network_state::RECENT_OUTBOUND_CONNECTIONS_MIN_DURATION;
-//use crate::peer_manager::network_state::RECONNECT_ATTEMPT_INTERVAL;
+use crate::peer_manager::network_state::RECONNECT_ATTEMPT_INTERVAL;
 use crate::peer_manager::peer_manager_actor::UPDATE_RECENT_OUTBOUND_CONNECTIONS_INTERVAL;
 use crate::peer_manager::testonly::start as start_pm;
 use crate::peer_manager::testonly::ActorHandler;
 use crate::tcp;
 use crate::testonly::make_rng;
 use crate::time;
-//use core::time::Duration;
 use near_o11y::testonly::init_test_logger;
 use near_primitives::network::PeerId;
 use near_store::db::TestDB;
 use std::sync::Arc;
-//use std::thread::sleep;
 
 async fn check_recent_outbound_connections(pm: &ActorHandler, wanted: Vec<PeerId>) {
     let got: Vec<PeerId> = pm
@@ -127,7 +125,7 @@ async fn test_outbound_connection_storage_order() {
     clock.advance(UPDATE_RECENT_OUTBOUND_CONNECTIONS_INTERVAL);
     check_recent_outbound_connections(&pm0, vec![id3.clone(), id2.clone(), id1.clone()]).await;
 
-    // disconnect pm2 and see that it remains in storage but moves to the end of the order
+    // disconnect pm2 and check that it remains in storage but moves to the end of the order
     pm2.disconnect(&id0).await;
     pm0.wait_for_num_connected_peers(2).await;
     clock.advance(UPDATE_RECENT_OUTBOUND_CONNECTIONS_INTERVAL);
@@ -162,9 +160,8 @@ async fn test_reconnect_after_restart_outbound_side() {
     // check that pm0 stores the outbound connection to pm1
     check_recent_outbound_connections(&pm0, vec![id1.clone()]).await;
 
+    // drop pm0 and start it again with the same db, check that it reconnects
     drop(pm0);
-
-    // start a new pm0 with the same db and see that it reestablishes the connection to pm1
     let pm0 = start_pm(clock.clock(), pm0_db, chain.make_config(rng), chain.clone()).await;
     pm0.wait_for_direct_connection(id1.clone()).await;
 
@@ -172,7 +169,7 @@ async fn test_reconnect_after_restart_outbound_side() {
     drop(pm1);
 }
 
-/*#[tokio::test]
+#[tokio::test]
 async fn test_reconnect_after_restart_inbound_side() {
     init_test_logger();
     let mut rng = make_rng(921853233);
@@ -197,17 +194,18 @@ async fn test_reconnect_after_restart_inbound_side() {
     // check that pm0 stores the outbound connection to pm1
     check_recent_outbound_connections(&pm0, vec![id1.clone()]).await;
 
-    //drop(pm1);
-    pm1.disconnect(&id0);
+    // have pm1 simply stop the connection and check that pm0 reconnects
+    pm1.disconnect(&id0).await;
     pm0.wait_for_num_connected_peers(0).await;
-
-    // start a new pm1 with the same config and see that pm0 reconnects to it
-    //let pm1 = start_pm(clock.clock(), TestDB::new(), pm1_cfg.clone(), chain.clone()).await;
-
     clock.advance(RECONNECT_ATTEMPT_INTERVAL);
+    pm0.wait_for_direct_connection(id1.clone()).await;
 
+    // drop pm1 and start it again with the same config, check that pm0 reconnects
+    drop(pm1);
+    let pm1 = start_pm(clock.clock(), TestDB::new(), pm1_cfg.clone(), chain.clone()).await;
+    clock.advance(RECONNECT_ATTEMPT_INTERVAL);
     pm0.wait_for_direct_connection(id1.clone()).await;
 
     drop(pm0);
     drop(pm1);
-}*/
+}
